Move the ivf muxer to a subcrate
