+++ Cargo.toml
@@ -12,7 +12,7 @@ autobins = false
 comparative_bench = ["aom"]
 decode_test = ["bindgen", "aom"]
 decode_test_dav1d = ["dav1d-sys"]
-binaries = ["y4m", "clap", "scan_fmt"]
+binaries = ["ivf", "y4m", "clap", "scan_fmt"]
 default = ["binaries", "nasm"]
 aom = ["cmake"]
 nasm = ["nasm-rs"]
@@ -30,6 +30,7 @@ num-traits = "0.2"
 paste = "0.1"
 dav1d-sys = { version = "0.1.2", optional = true }
 scan_fmt = { version = "0.1.3", optional = true }
+ivf = { path = "ivf/", optional = true }
 
 [build-dependencies]
 cmake = { version = "0.1", optional = true }
@@ -64,3 +65,6 @@ codegen-units = 1  # if > 1 enables parallel code generation which improves
 
 [profile.bench]
 codegen-units = 1
+
+[workspace]
+members = ["ivf"]
