+++ src/bin/muxer.rs
@@ -7,37 +7,11 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
-use bitstream_io::{BitWriter, LittleEndian};
 use crate::decoder::VideoDetails;
-use std::io;
 use std::io::Write;
 use std::slice;
 
-pub fn write_ivf_header(
-  output_file: &mut dyn io::Write, width: usize, height: usize, num: usize,
-  den: usize
-) {
-  let mut bw = BitWriter::endian(output_file, LittleEndian);
-  bw.write_bytes(b"DKIF").unwrap();
-  bw.write(16, 0).unwrap(); // version
-  bw.write(16, 32).unwrap(); // version
-  bw.write_bytes(b"AV01").unwrap();
-  bw.write(16, width as u16).unwrap();
-  bw.write(16, height as u16).unwrap();
-  bw.write(32, num as u32).unwrap();
-  bw.write(32, den as u32).unwrap();
-  bw.write(32, 0).unwrap();
-  bw.write(32, 0).unwrap();
-}
-
-pub fn write_ivf_frame(
-  output_file: &mut dyn io::Write, pts: u64, data: &[u8]
-) {
-  let mut bw = BitWriter::endian(output_file, LittleEndian);
-  bw.write(32, data.len() as u32).unwrap();
-  bw.write(64, pts).unwrap();
-  bw.write_bytes(data).unwrap();
-}
+pub use ivf::*;
 
 pub fn write_y4m_frame(y4m_enc: &mut y4m::Encoder<'_, Box<dyn Write>>, rec: &rav1e::Frame, y4m_details: VideoDetails) {
   let pitch_y = if y4m_details.bit_depth > 8 { y4m_details.width * 2 } else { y4m_details.width };
